requirements.txt
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.10
gunicorn==21.2.0
python-decouple==3.8
django-cors-headers==4.3.1
setuptools>=69.0.0
